refactor(macos): split capture.rs into focused modules
